feat(motor): 支持浮点数控制和新电机类型
- 修改电机控制接口，支持浮点数输入
- 添加对DeepRobotics J60电机的支持
- 更新PID控制器的数据类型为浮点数
- 优化电机位置和速度环控制逻辑
- 改进上位机通信协议的可配置性
- 重构电机类继承结构，增强扩展性
- 修复DJI M3508电机CAN帧发送问题
- 完善电机数据获取接口，增加温度和电流读取

+++ motorPack/dji_m3508.cpp
@@ -45,13 +45,13 @@ void DjiM3508::SendData() const {
         if (id_ <= 4) {
             frame.id = DJI_3508_TX_ID_LOW;
             int idx = (id_ - 1) * 2;
-            frame.data[idx] = (target_current_ >> 8) & 0xFF;
-            frame.data[idx + 1] = target_current_ & 0xFF;
+            frame.data[idx] = ((int16_t)target_current_ >> 8) & 0xFF;
+            frame.data[idx + 1] = ((int16_t)target_current_ & 0xFF);
         } else {
             frame.id = DJI_3508_TX_ID_HIGH;
             int idx = (id_ - 5) * 2;
-            frame.data[idx] = (target_current_ >> 8) & 0xFF;
-            frame.data[idx + 1] = target_current_ & 0xFF;
+            frame.data[idx] = ((int16_t)target_current_ >> 8) & 0xFF;
+            frame.data[idx + 1] = ((int16_t)target_current_ & 0xFF);
         }
         (void)can_send(can_dev_, &frame, K_NO_WAIT, NULL, NULL);
     }
@@ -60,7 +60,7 @@ void DjiM3508::SendData() const {
 
 
 
-void DjiM3508::SetCurrentOpenLoop(int target) {
+void DjiM3508::SetCurrentOpenLoop(float target) {
     target_current_ = target;
 
     if (my_handler_&&sync_send_mode_)
